### 2018-10-07 #1001 @whedon
[PRE REVIEW]: Prest: Open-Source Software for Computational Revealed Preference Analysis
**Submitting author:** @georgiosgerasimou (<a href="http://orcid.org/0000-0003-3712-3154">Georgios Gerasimou</a>)
**Repository:** <a href="https://github.com/prestsoftware/prest" target ="_blank">https://github.com/prestsoftware/prest</a>
**Version:** v.0.9.7
