### 2018-10-07 #1001 @arfon
Hi @georgiosgerasimou - please take a look a the example paper in our submission guidelines and format the YAML header _exactly_ as shown in the example: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

(Note that YAML is whitespace sensitive)
