+++ paper/paper.md
@@ -7,8 +7,10 @@ tags:
   - simulations
 authors:
   - name: Georgios Gerasimou
+    orcid: 0000-0003-3712-3154
   - affiliation: 1
   - Matúš Tejiščák
+    orcid: 0000-0003-0000-0000
   - affiliation: 2
 affiliations:
   - name: School of Economics & Finance, University of St Andrews
