### 2018-10-07 #1001 @georgiosgerasimou
The updated paper.md file has now been uploaded to the repository.
