### 2018-10-07 #1001 @arfon
Looks like it's still broken. You can test changes yourself here @georgiosgerasimou with the command `@whedon generate pdf`
