### 2018-10-07 e86be4c
More YAML fixing.
