+++ paper/paper.bib
@@ -265,3 +265,32 @@
 	keywords =     {},
 	source =       {},
 }
+
+@ARTICLE{heufer-hjertstrand15,
+	AUTHOR =       {Jan Heufer and Per Hjertstrand},
+	TITLE =        {Consistent Subsets: Computationally Feasible Methods to Compute the {H}outman-{M}aks-Index},
+	JOURNAL =      {Economics Letters},
+	YEAR =         {2015},
+	volume =       {128},
+	pages =        {87-89},
+	month =        {},
+	note =         {},
+	abstract =     {},
+	keywords =     {},
+	source =       {},
+}
+
+@ARTICLE{beatty-crawford11,
+	AUTHOR =       {Timothy K. M. Beatty and Ian A. Crawford},
+	TITLE =        {How Demanding is the Revealed Preference Approach to Demand},
+	JOURNAL =      {American Economic Review},
+	YEAR =         {2011},
+	volume =       {101},
+	number =       {6},
+	pages =        {2782-2795},
+	month =        {},
+	note =         {},
+	abstract =     {},
+	keywords =     {},
+	source =       {},
+}
