### 2018-10-07 eec6b34
Updated references
