+++ paper/paper.bib
@@ -294,3 +294,18 @@
 	keywords =     {},
 	source =       {},
 }
+
+@ARTICLE{dean-martin16,
+	AUTHOR =       {Mark Dean and Daniel Martin},
+	TITLE =        {Measuring Rationality with the Minimum Cost of Revealed Preference Violations},
+	JOURNAL =      {Review of Economics and Statistics},
+	YEAR =         {2016},
+	volume =       {98},
+	number =       {3},
+	pages =        {524-534},
+	month =        {},
+	note =         {},
+	abstract =     {},
+	keywords =     {},
+	source =       {},
+}
