### 2018-10-09 #1001 @HaoZeke
I would be glad to look this over.
