### 2018-10-14 #1015 @whedon
[REVIEW]: Prest: Open-Source Software for Computational Revealed Preference Analysis
**Submitting author:** @georgiosgerasimou (<a href="http://orcid.org/0000-0003-3712-3154">Georgios Gerasimou</a>)
**Repository:** <a href="https://github.com/prestsoftware/prest" target ="_blank">https://github.com/prestsoftware/prest</a>
**Version:** v.0.9.7

- [x] **Version:** Does the release version given match the GitHub release (v.0.9.7)?
