### 2018-10-14 ad9f65f
paper: Update bibliography with betterbib
