+++ paper/paper.bib
@@ -1,311 +1,377 @@
-@ARTICLE{richter66,
-  AUTHOR =       {Marcel K. Richter},
-  TITLE =        {Revealed Preference Theory},
-  JOURNAL =      {Econometrica},
-  YEAR =         {1966},
-  volume =       {34},
-  number =       {3},
-  pages =        {635-645},
-  month =        {},
-  note =         {},
-  abstract =     {},
-  keywords =     {},
-  source =       {},
+%comment{This file was created with betterbib v3.4.0.}
+
+
+@article{richter66,
+ author = {Richter, Marcel K.},
+ title = {Revealed Preference Theory},
+ journal = {Econometrica},
+ year = {1966},
+ volume = {34},
+ number = {3},
+ pages = {635},
+ month = jul,
+ note = {},
+ abstract = {},
+ keywords = {},
+ source = {Crossref},
+ doi = {10.2307/1909773},
+ url = {https://doi.org/10.2307/1909773},
+ publisher = {JSTOR},
+ issn = {0012-9682},
 }
 
-@ARTICLE{arrow59,
-  AUTHOR =       {Kenneth J. Arrow},
-  TITLE =        {Rational Choice Functions and Orderings},
-  JOURNAL =      {Economica},
-  YEAR =         {1959},
-  volume =       {26},
-  number =       {102},
-  pages =        {121-127},
-  month =        {},
-  note =         {},
-  abstract =     {},
-  keywords =     {},
-  source =       {},
+@article{arrow59,
+ author = {Arrow, Kenneth J.},
+ title = {Rational Choice Functions and Orderings},
+ journal = {Economica},
+ year = {1959},
+ volume = {26},
+ number = {102},
+ pages = {121},
+ month = may,
+ note = {},
+ abstract = {},
+ keywords = {},
+ source = {Crossref},
+ doi = {10.2307/2550390},
+ url = {https://doi.org/10.2307/2550390},
+ publisher = {JSTOR},
+ issn = {0013-0427},
 }
 
-
-@ARTICLE{samuelson38,
-  AUTHOR =       {Paul A. Samuelson},
-  TITLE =        {A Note on the Pure Theory of Consumer's Behaviour},
-  JOURNAL =      {Economica},
-  YEAR =         {1938},
-  volume =       {5},
-  number =       {17},
-  pages =        {61-71},
-  month =        {},
-  note =         {},
-  abstract =     {},
-  keywords =     {},
-  source =       {},
+@article{samuelson38,
+ author = {Samuelson, P. A.},
+ title = {A Note on the Pure Theory of Consumer's Behaviour},
+ journal = {Economica},
+ year = {1938},
+ volume = {5},
+ number = {17},
+ pages = {61},
+ month = feb,
+ note = {},
+ abstract = {},
+ keywords = {},
+ source = {Crossref},
+ doi = {10.2307/2548836},
+ url = {https://doi.org/10.2307/2548836},
+ publisher = {JSTOR},
+ issn = {0013-0427},
 }
 
-@ARTICLE{bewley02,
-  AUTHOR =       {Truman F. Bewley},
-  TITLE =        {Knightian Decision Theory. {P}art {I}},
-  JOURNAL =      {Decisions in Economics and Finance},
-  YEAR =         {2002},
-  volume =       {25},
-  number =       {2},
-  pages =        {79-110},
-  month =        {},
-  note =         {Working-paper version: 1986},
-  abstract =     {},
-  keywords =     {},
-  source =       {},
+@article{bewley02,
+ author = {Bewley, Truman F.},
+ title = {Knightian decision theory. Part I},
+ journal = {Decisions in Economics and Finance},
+ year = {2002},
+ volume = {25},
[…]
