### 2018-10-16 ecfd98a
readme: Installation updates
Specifically, this clarifies `pip` usage in a virtual environment.
