+++ README.md
@@ -22,17 +22,20 @@ Dependencies:
 * Rust 1.26 stable + Cargo
 * Python 3 (we use 3.6 and 3.7) + pip
 
+Install the dependencies, compile everything and run:
+
+```bash
+$ pip install --user -r gui/requirements.txt
+$ make run
+```
+
 Optionally, work in a virtual environment:
 
 ```bash
 $ python3 -m venv prest.env
 $ source prest.env/bin/activate
-```
-
-Install the dependencies, compile everything and run:
-
-```bash
-$ pip install --user -r gui/requirements.txt
+# The next line is different from the previous installation command.
+$ pip install -r gui/requirements.txt
 $ make run
 ```
 
