### 2018-10-16 prestsoftware/prest#3 @HaoZeke
Installation hurdles
# Dependencies and over specified defaults
**Acceptance blocker (https://github.com/openjournals/joss-reviews/issues/1015).**

The default installation stalls without `mypy` which is really meant for tests and should be decoupled from the main installation procedure or documented as a requirement.

```bash
make run
```

In fact, as it stands, the `make run` command also builds documentation, which also should be optional.

The error generated by not having `mypy` is:

``` bash
mypy --platform win32 --ignore-missing-imports --check-untyped-defs --strict-optional main.py
make[1]: execvp: mypy: Permission denied
make[1]: *** [Makefile:7: .typecheck-ts] Error 127
make[1]: Leaving directory '/home/haozeke/Git/Github/Reviews/prest/gui'
make: *** [Makefile:16: gui/.typecheck-ts] Error 2
```

# Warnings and Cosmetics

Also, the documentation generation has several warnings of the form:

`/home/haozeke/Git/Github/Reviews/prest/docs/src/references.rst: WARNING: Citation [redelmeier-shafir95] is not referenced.`

If these are not intentional these should be either suppressed or fixed.

In addition to this there are some minor changes which I am adding as pull requests, namely:

- The installation instructions create a `venv` which should be ignored
- The installation instructions differ for the virtual environment case (`pip install -r` instead of `pip install --user -r`
