### 2018-10-16 #1015 @HaoZeke
@georgiosgerasimou I have raised some concerns regarding the installation procedure in https://github.com/prestsoftware/prest/issues/3.

Since there are multiple possible ways of resolving them, please discuss them on the appropriate issue.
