### 2018-10-16 prestsoftware/prest#3 @ziman
Thank you for catching this!

Indeed, `mypy` is a build requirement, which I did not realise. My view is that typechecking is an integral part of the build, and therefore I will add `mypy` to `requirements.txt`.

The documentation is also part of the build — it can be invoked from the GUI so it must be built before you run Prest. For distribution, the whole HTML directory is also embedded in the distribution executable.

We should probably document this as well. Thank you for pointing this out!
