### 2018-10-16 prestsoftware/prest#3 @ziman
I also agree that the doc build warnings should be suppressed (we include unreferenced citations intentionally).
