### 2018-10-17 f45f6c3
Add mypy to requirements.txt.
