+++ gui/requirements.txt
@@ -1,3 +1,4 @@
+mypy
 hypothesis
 numpy
 openpyxl
