### 2018-10-17 92c4c9d
Mention doc build + typechecking in README.
