### 2018-10-18 prestsoftware/prest#3 @HaoZeke
Thanks for resolving these (warnings can be handled later) in 92c4c9d41e0dcd1a191a6faa377796e8dfa3d75a and f45f6c36ff2b54d08d6a949f9593c47e314127ef!
