### 2018-10-18 prestsoftware/prest#5 @HaoZeke
Functionality Test Suite
**Acceptance Blocker (https://github.com/openjournals/joss-reviews/issues/1015)**

# Details
As per the JOSS review guidelines, the tests to be included are to be classified as:

> Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.

    Good: An automated test suite hooked up to an external service such as Travis-CI or similar
    OK: Documented manual steps that can be followed to check the expected functionality of the software (e.g. a sample input file to assert behaviour)
    Bad (not acceptable): No way for you the reviewer to check whether the software works

At the moment, in-spite of very well documented functionality and some example inputs, I cannot seem to find any overall tests which showcase high level functionality in a deterministic manner.

This is somewhat related to the issue of examples, however after reading the documentation I determined that the sample inputs and documentation provide appropriate example usage.
