### 2018-10-18 19ac7f4
Update CONTRIBUTING.
