### 2018-10-18 prestsoftware/prest#5 @ziman
The [test suite documented in the README](https://github.com/prestsoftware/prest#testing) includes an [integration test](https://github.com/prestsoftware/prest/blob/master/gui/test/integration_test.py), which runs the whole pipeline including the Rust core on the examples provided in the documentation and compares the outputs with the expected values. I think that this test suite covers the main subset of the high-level functionality of the software reasonably but perhaps my reading of the guidelines is incorrect.

What other tests would you prefer to see? Or perhaps we should document the tests better?
