### 2018-10-18 prestsoftware/prest#5 @HaoZeke
Thanks for highlighting that! I think some slight changes to the documentation (the generated html files) which could like to the integration tests would be a good idea.

Having gone through the code and the test itself I am closing this here, however please consider updating the `html` docs.
