### 2018-10-18 prestsoftware/prest#5 @HaoZeke
In particular, I was wondering if perhaps the CLI tool mentioned in the paper (to model estimation for testing and benchmarking) would be a part of this repo eventually.
