### 2018-10-18 #1015 @HaoZeke
@arfon This is now fully compliant with the JOSS guidelines IMO.

@georgiosgerasimou and @ziman, thank you so much for being prompt and accommodative to my many queries and clarifications. It's an amazing software, with great docs and code architecture, and I am glad I got the chance to review it! Also, thanks for taking the time to ensure JOSS compliance.
