### 2018-10-19 prestsoftware/prest#5 @ziman
The fact that you could not find the test indicates that it indeed needs to be documented properly.

However, the HTML documentation is intended mainly for end users and I feel like developer information, such as test descriptions, would be better placed elsewhere, such as CONTRIBUTING.md or perhaps README.md.

As for [the CLI tool](https://github.com/prestsoftware/prest/blob/master/gui/cli.py), it is indeed silently included in the repository. We do not document it as supported functionality because we still need to find a way to make model selection and export variant selection more ergonomic — at the moment you are required to select them by their description string, exactly character by character.
