### 2018-10-19 81a0308
Updated the CCGT16 reference with more details
