### 2018-10-19 af076e3
Updated the CCGT16 reference with more details
