+++ paper/paper.bib
@@ -189,7 +189,7 @@
 @article{CCGT16,
  author = {Costa-Gomes, Miguel and Cueva, Carlos and Gerasimou, Georgios and Tejiščák, Matúš},
  title = {Choice, Deferral and Consistency},
- journal = {School of Economics & Finance Discussion Paper 1416, University of St Andrews},
+ journal = {School of Economics \& Finance Discussion Paper 1416, University of St Andrews},
  year = {2016},
  volume = {},
  number = {},
