### 2018-10-20 85ed7bc
Add v0.9.8 + v0.9.9 to changelog.
