### 2018-10-20 561eb92
Bump core version to 0.9.9.
