### 2018-10-20 #1015 @arfon
I'm going to try accepting this automatically here (this is new fucntionality for Whedon). If this goes wrong I will have to fix this up tomorrow evening my time (currently typing this from my phone) 😬
