### 2018-10-20 #1015 @arfon
@HaoZeke - many thanks for your review here! ✨
If
@georgiosgerasimou - your paper is now accepted into JOSS and your DOI is
https://doi.org/10.21105/joss.01015 ⚡️🚀💥
